requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.6
django-crispy-forms==1.11.0
django-environ==0.4.5
django-heroku==0.3.1
djangorestframework==3.12.2
environ==1.0
gunicorn==20.0.4
psycopg2==2.8.6
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
